Restructure ViewModel Folder
