+++ budget-binder-multiplatform-app/src/jvmMain/kotlin/de/hsfl/budgetBinder/screens/login/LoginComponent.kt
@@ -14,8 +14,8 @@ import de.hsfl.budgetBinder.di
 import de.hsfl.budgetBinder.compose.dialog.ServerUrlDialog
 import de.hsfl.budgetBinder.compose.icon.AppIcon
 import de.hsfl.budgetBinder.compose.textfield.EmailTextField
-import de.hsfl.budgetBinder.presentation.login.LoginViewModel
-import de.hsfl.budgetBinder.presentation.login.LoginEvent
+import de.hsfl.budgetBinder.presentation.viewmodel.login.LoginViewModel
+import de.hsfl.budgetBinder.presentation.viewmodel.login.LoginEvent
 import kotlinx.coroutines.flow.collectLatest
 import org.kodein.di.instance
