+++ budget-binder-multiplatform-app/src/jvmMain/kotlin/de/hsfl/budgetBinder/screens/register/RegisterComponent.kt
@@ -12,8 +12,8 @@ import androidx.compose.ui.unit.dp
 import de.hsfl.budgetBinder.compose.icon.AppIcon
 import de.hsfl.budgetBinder.di
 import de.hsfl.budgetBinder.compose.textfield.EmailTextField
-import de.hsfl.budgetBinder.presentation.register.RegisterEvent
-import de.hsfl.budgetBinder.presentation.register.RegisterViewModel
+import de.hsfl.budgetBinder.presentation.viewmodel.register.RegisterEvent
+import de.hsfl.budgetBinder.presentation.viewmodel.register.RegisterViewModel
 import kotlinx.coroutines.flow.collectLatest
 import org.kodein.di.instance
